Organize centrality and traversal
